[RW-32] Add localization helper for collated sort and function to get a system path from an alias
